www-client/nyxt: "working" dep versions, build almost works; WIP
All the dependencies should be in the tree now but SBCL cannot save the
compiled image of Nyxt because of a multi-threading issue which I have
yet to solve. The versions of each lisp package are roughly pinned at
the same places as they are by quicklisp or Nyxt, but I've taken the
liberty to use the newer versions of some. Also there may be missing
runtime dependencies or unlisted dependencies for certain ebuilds that
happen to be installed by another required package. A thorough review of
this is still needed.